+++ cdk_framework/lib/config/testcase_config/basic-test-case.yml
@@ -1,8 +1,7 @@
 ---
 testcase:
   clusterName: amdCluster
-  sampleAppImageURL: "public.ecr.aws/aws-otel-test/aws-otel-java-spark:latest"
-  sampleAppMode: push
+  baseScenario: "otlp"
   aocConfig:
     extensions:
       health_check:
