+++ cdk_framework/lib/resource_constructs/otlp_test_case_constructs/aoc-config-map-construct.ts
@@ -1,6 +1,6 @@
 import { Construct } from 'constructs';
 import { Cluster, FargateCluster } from 'aws-cdk-lib/aws-eks';
-import { NamespaceConstruct } from './namespace-construct';
+import { NamespaceConstruct } from '../universal_constructs/namespace-construct';
 
 
 export class AOCConfigMapConstruct extends Construct {
