+++ cdk_framework/lib/resource_constructs/otlp_test_case_constructs/otlp-aoc-deployment-construct.ts
@@ -1,13 +1,13 @@
 import { Construct } from 'constructs';
 import { Cluster, FargateCluster } from 'aws-cdk-lib/aws-eks';
 import { AOCConfigMapConstruct } from './aoc-config-map-construct';
-import { NamespaceConstruct } from './namespace-construct';
-import { ServiceAccountConstruct } from './service-account-construct';
+import { NamespaceConstruct } from '../universal_constructs/namespace-construct';
+import { ServiceAccountConstruct } from '../universal_constructs/service-account-construct';
 
-export class AOCDeploymentConstruct extends Construct {
+export class OTLPAOCDeploymentConstruct extends Construct {
     aocDeployment: Construct
 
-    constructor(scope: Construct, id: string, props: AOCDeploymentConstructProps) {
+    constructor(scope: Construct, id: string, props: OTLPAOCDeploymentConstructProps) {
         super(scope, id);
 
         const aocConfigMountPath = '/aoc'
@@ -126,12 +126,13 @@ export class AOCDeploymentConstruct extends Construct {
         
         this.aocDeployment = props.cluster.addManifest('aoc-deployment', aocDeploymentManifest)
         this.aocDeployment.node.addDependency(props.namespaceConstruct.namespace)
+        // Are these dependencies needed?
         this.aocDeployment.node.addDependency(props.serviceAccountConstruct.serviceAccount)
         this.aocDeployment.node.addDependency(props.aocConfigMapConstruct.aocConfigMap)
     }
 }
 
-export interface AOCDeploymentConstructProps {
+export interface OTLPAOCDeploymentConstructProps {
     cluster: Cluster | FargateCluster
     namespaceConstruct: NamespaceConstruct
     aocAppLabel: string
