+++ cdk_framework/lib/resource_constructs/prometheus_test_case_constructs/pull-mode-sample-app-construct.ts
@@ -1,6 +1,6 @@
 import { Construct } from 'constructs';
 import { Cluster, FargateCluster } from 'aws-cdk-lib/aws-eks';
-import { NamespaceConstruct } from './namespace-construct';
+import { NamespaceConstruct } from '../universal_constructs/namespace-construct';
 
 
 export class PullModeSampleAppDeploymentConstruct extends Construct {
@@ -24,14 +24,14 @@ export class PullModeSampleAppDeploymentConstruct extends Construct {
 
                 selector: {
                     matchLabels: {
-                        app: props.sampleAppLabelSelector
+                        app: props.sampleAppLabel
                     }
                 },
 
                 template: {
                     metadata: {
                         labels: {
-                            app: props.sampleAppLabelSelector
+                            app: props.sampleAppLabel
                         }
                     },
 
@@ -85,8 +85,7 @@ export class PullModeSampleAppDeploymentConstruct extends Construct {
 export interface PullModeSampleAppDeploymentConstructProps {
     cluster: Cluster | FargateCluster
     namespaceConstruct: NamespaceConstruct
-    sampleAppLabelSelector: string
-    sampleAppImageURL: string
+    sampleAppLabel: string
     listenAddressHost: string
     listenAddressPort: number
     region: string
