+++ cdk_framework/lib/resource_constructs/unimplemented_constructs/cluster-role-binding-construct.ts
@@ -1,6 +1,6 @@
 import { Construct } from 'constructs';
 import { Cluster, FargateCluster } from 'aws-cdk-lib/aws-eks';
-import { NamespaceConstruct } from '../basic_constructs/namespace-construct';
+import { NamespaceConstruct } from '../universal_constructs/namespace-construct';
 
 export class ClusterRoleBindingConstruct extends Construct{
     constructor(scope: Construct, id: string, props: ClusterRoleBindingConstructProps){
