+++ cdk_framework/lib/resource_constructs/unimplemented_constructs/logs-sample-fargate-deploy-construct.ts
@@ -1,7 +1,6 @@
 import { Construct } from 'constructs';
 import { Cluster, FargateCluster } from 'aws-cdk-lib/aws-eks';
-import { NamespaceConstruct } from '../basic_constructs/namespace-construct';
-
+import { NamespaceConstruct } from '../universal_constructs/namespace-construct';
 
 export class LogsSampleFargateDeployConstruct extends Construct{
     constructor(scope: Construct, id: string, props: LogsSampleFargateDeployConstructProps){
