+++ cdk_framework/lib/resource_constructs/universal_constructs/namespace-construct.ts
